requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyJWT
